finished reponsiveness homepage
after fixing the reponsiveness for phone and similar items. I had to  continue off with my idea of having tabs on the right of my qualifications. Using the flexgrid setup made by my teacher, I got it to work but not in the way I wanted to sadly. Now it works after a small tear in my eye.

I also made some small additional changes such as increasing the button size, font size and picture centering.