github repo list
after working on this feature way too much, I have managed to create a link between my github repositories and the website. My website will now gather my repository name, desc, lastupdate date, and last commit message. This will be displayed in a div. All the data is saved in cookies that will refresh every 15 minutes in order to keep anonymous requests low and not get blocked out. I have had 3 total reworks of this project and are visible in the repository