project page styling
I have started out by reconsidering all my hard work to show repo latest commit messages. It will, at some point, be too many requests and just not worth it anymore. After removing that piece of code, The styling was added for the projects page and a nice background which I still need to do properly because there is an better way.