small fixes
did some more website testing and found out my footer would have been an issue. as such I removed all styling on it and just left it there to indicate that all the parts in it are part of the footer nav. I also added a link to the repositories listed and added the repos from my school account