updated animation for phones
After wanting to start working on responsiveness I walked into a issue with my current animation. It did NOT look good on phones. So I had to alter it. The way I decided to go was to make a seperate component with about the same elements and code, but slightly adjusted so it would work with phones and still look nice without being too different. I think it worked out well with minimal work and Im not sure but I would say its still performance-efficient