Navigation and scroll fix
As I started working on the navigation I was delighted to see that most of the functions I already had fitted very well with the navigation. The only thing I had to do was to take out some variables from the function that handled scrolling(big mistake...). This ended up being a huge issue because the function was completely isolated because of the fact that it went to a child and back. It was snapshotted before all the usestates, which were needed for a reason I can not remeber, were filled. This caused a huge issue and a massive headache. My solution I ended up with was to make the function that handles the scrolling depedent on a boolean that it could read and the function given to the child could alter. This indeed ended up working and although it was a huge headache, it helped me understand my code more and help reduce clutter code.