finalised desktop nav and started working on threejs
I have been stuck for two days on my navbar and how to do the coloring properly. I kept trying to make the parent call a child method. Then it came to me, instead of going against the water, why not flow with it? As I realised that I moved the method for the navcolors up and passed it onto the child element. This way the parent could make use of it and so could the child. This will also help with the mobile variant of the nav because (i do not know yet) it will be using the same code. After that I started reading up on threejs documentation as it has been a while since I used it and instantly got stuck on reacts strictmode. It was double rendering and causing issues because it kept creating to canvasses. After fighting with it for a while and not wanting to be defeated by strictmode I could instead of deleting, which didnt work by the way, just replace said child. This solved the issue and now i am happily rendering a cube. I will be going ahead with the rendering part for threejs for the foreesable future