big update landing and projects
i have again forgotten to commit. a bad habit of mine. Anyway alot has changed since last commit. I have been able to finish the threejs rendering on my website, although I am not that happy with it. It could have been MUCH MUCH better but due to timeconstraints and the fact my mind was having a hard time wrapping around what is happening, it ended up like this. Its still not bad at all for my first time really working with it. I have also, after ALOT of work been able to import my own model. After that I started working on the about section. This was by far the easiest after all the hard work from before. I am happy something went smooth for once. Now im working and basically finished with the projects spotlight section. it is functionally done but I want to make a link to another page with more projects on them. that should look better and be easier to view and work with. my biggest issues have been threejs working weird and trying to put it in a ts file. I never really worked with TS and wanted to try it out. its a bit messy but not too bad