section scrolling fix
It took me a bit and alot of concepts. What it came down to was that the scrolling on trackpad was causing a huge issue where it would fire a million times a second. This would cause issues when you only have three pages. I was trying to do a whole workaround by using the onscoll listener. This was not really the greatest idea because it would lack alot of info to properly scroll. Later on, after doing research I found out how to properly throttle scrolls so it doesnt go to a million. This made me realise I could use the old setup with just the throttle.