the final triumph: responsiveness and more pages
 I have finally, what I think, finished the second variant of my cv. It has everything I could fit within a week of building and my knowledge that i have built up over the year and a half of progrmamming. I wouldnt say im super proud of it but it is definitely something I would like to present to other people as my work. It is functional and , I hope, properly accessible to people.

now onto whats added. I reworked alot of the existing css to work with mobile. I know alot of people like to do it the other way around, i dont. but after this I might. I was having, and left in, alot of issues with proper image placing. it would unneccesarily take up space it did not deserve for its size. now while this sort of worked out there might still be some parts that have it.

I added react routing to even get to the projects page. this wasnt hard as i did it before.

I had to rework the existing nav to work with mobile. I originally intended to remake the whole thing, hence why the component is called Desktopnav. I found it uneccessary and just added cool special styling. I am quite proud of that.

that is all I can remember. Most of the issues have just been the css not working properly with my mind. Its something I keep having issues with and, at this point, just have to relearn. but thats a story for another day. now enjoy the portfoliopage I made :)